Refactor fast RNN
